==> 01- Pull Reddit Data.ipynb <==
### This notebook utilizes the pushshift API to pull data about posts in the r/food submreddit

import requests
import pandas as pd
import time
import datetime as dt

base_url='https://api.pushshift.io/reddit/search/submission/?'

utc_2_weeks = 1653077324

def reddit_posts(topic):
    res = requests.get(base_url, params={'subreddit':topic,
                                         'size':100
                                         #,'selftext':True
                                         ,'before': utc_2_weeks
                                        })
    if res.status_code ==200:
        data = res.json()
        df = pd.DataFrame(data['data'])[['id','subreddit', 'title', 'selftext','created_utc','num_comments','score']]
        return df #, time_stamp
    else:
        print('request failed')

def reddit_posts1(topic, time_stamp):
    res = requests.get(base_url, params={'subreddit':topic,
                                         'size':100
                                         ,'before':time_stamp
                                         #,'selftext':True
                                        })
    if res.status_code==200:
        data = res.json()
        df = pd.DataFrame(data['data'])[['id','subreddit', 'title', 'selftext','created_utc','num_comments','score']]
        return df #, time_stamp
    else:
        print('request failed')

posts=[]
data1 = reddit_posts('Food')
posts.append(data1)
time_stamp = data1[-1:]['created_utc']
time.sleep(1)
for x in range(250):
    data1 = reddit_posts1('Food', time_stamp)
    posts.append(data1)
    time_stamp = data1[-1:]['created_utc']
    time.sleep(1)
reddit_posts_df = pd.concat(posts)

reddit_posts_df.head()

reddit_posts_df['score'].value_counts()

reddit_posts_df['num_comments'].value_counts()

reddit_posts_df.sort_values(by='score', ascending=False)

Taking a quick look at these scores and comparing to the scores seen when visiting the the webpage it is clear that they are inaccurate. See the notebook 'redditAPI' for how to pull the accurate data correctly.

reddit_posts_df[-1:]['created_utc']

utc_2_weeks = 1653077324

def reddit_posts2(topic):
    res = requests.get(base_url, params={'subreddit':topic,
                                         'size':100
                                         #,'selftext':True
                                         ,'before': reddit_posts_df[-1:]['created_utc']
                                        })
    if res.status_code ==200:
        data = res.json()
        df = pd.DataFrame(data['data'])[['id','subreddit', 'title', 'selftext','created_utc','num_comments','score']]
        return df #, time_stamp
    else:
        print('request failed')

def reddit_posts1(topic, time_stamp):
    res = requests.get(base_url, params={'subreddit':topic,
                                         'size':100
                                         ,'before':time_stamp
                                         #,'selftext':True
                                        })
    if res.status_code==200:
        data = res.json()
        df = pd.DataFrame(data['data'])[['id','subreddit', 'title', 'selftext','created_utc','num_comments','score']]
        return df #, time_stamp
    else:
        print('request failed')

posts=[]
data1 = reddit_posts2('Food')
posts.append(data1)
time_stamp = data1[-1:]['created_utc']
time.sleep(1)
for x in range(250):
    data1 = reddit_posts1('Food', time_stamp)
    posts.append(data1)
    time_stamp = data1[-1:]['created_utc']
    time.sleep(1)
reddit_posts_df_2 = pd.concat(posts)

reddit_posts_df_2

reddit_posts_df_2[-1:]['created_utc']

reddit_df = pd.concat([reddit_posts_df, reddit_posts_df_2])

len(reddit_df['id'].unique()) #confirming pulls were unique

reddit_df['score'].value_counts()

reddit_df.to_csv('./food_trends/Data/reddit_6_months.csv', index=False)
