==> 02- redditAPI.ipynb <==
### Reddit API

The information pulled from the pushshift API included inaccurate values for post scores and post number of comments. This notebooks uses the Reddit API to pull in the correct scores and number of comments for the post ids sourced earlier

# pip install praw

import requests
import pandas as pd
import time
import datetime as dt
import praw

To get these credentials I applied for an api key and ran curl commands in my terminal(CLI)

reddit = praw.Reddit(
    client_id="yEQokwxZEpfab7-hjZGYqQ",
    client_secret="0wfgItj8y8K12zwbKGBWEQantbeK8A",
    password="capstone22",
    user_agent="food trends 2 by u/ekfuller",
    username="ekfuller",
)

If the following cell runs then the credentials are active and valid

for submission in reddit.subreddit("food").hot(limit=25):
    print(submission.id, submission.score, submission.num_comments)

# read in reddit data from csv to get post ids

posts_df = pd.read_csv('../food_trends/Data/reddit_6_months.csv')

posts_df.head(2)

posts_df.drop(columns=['subreddit','selftext', 'num_comments','score'], inplace=True)

posts_df

Using the post ids, pull the score and number of comments per post. It gets broken into 100 post chunks to comply with pull limits. I further broke the data into groups of 1000 since the function takes a long time to run and it was easier to save in chunks and then recombine

def get_scores(posts):
    for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return posts

def get_lists(chunks):
    list_of_chunks = []
    for x in range(0,len(chunks)):
        list_of_chunks.append(list(chunks[x]))
    return list_of_chunks

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_10 = posts_df[:1000]

chunks = [posts_df_10['id'][x:x+100] for x in range(0, len(posts_df_10), 100)]

increments_10 = get_lists(chunks)

get_posts(increments_10)

scores_df_10 = pd.DataFrame([ids, scores, comments]).T

scores_df_10.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_10.to_csv('./Data/0_to_10k_scores.csv', index=False)

posts_df_20 = posts_df[1000:2000]

chunks = [posts_df_20['id'][x:x+100] for x in range(0, len(posts_df_20), 100)]

increments_20 = get_lists(chunks)

get_posts(increments_20)

scores_df_20 = pd.DataFrame([ids, scores, comments]).T

scores_df_20.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_20

scores_df_20.to_csv('./Data/10_to_20k_scores.csv', index=False)

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_30 = posts_df[2000:3000]

chunks = [posts_df_30['id'][x:x+100] for x in range(0, len(posts_df_30), 100)]

increments_30 = get_lists(chunks)

get_posts(increments_30)

scores_df_30 = pd.DataFrame([ids, scores, comments]).T

scores_df_30.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_30.to_csv('./Data/2_to_3k_scores.csv', index=False)

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_40 = posts_df[3000:4000]

chunks = [posts_df_40['id'][x:x+100] for x in range(0, len(posts_df_40), 100)]

increments_40 = get_lists(chunks)

get_posts(increments_40)

scores_df_40 = pd.DataFrame([ids, scores, comments]).T

scores_df_40.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)

scores_df_40.to_csv('./Data/3_to_4k_scores.csv', index=False)

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_50 = posts_df[4000:5000]

chunks = [posts_df_50['id'][x:x+100] for x in range(0, len(posts_df_50), 100)]

increments_50 = get_lists(chunks)

get_posts(increments_50)

scores_df_50 = pd.DataFrame([ids, scores, comments]).T

scores_df_50.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_50.to_csv('./Data/4_to_5k_scores.csv', index=False)

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_60 = posts_df[5000:6000]

chunks = [posts_df_60['id'][x:x+100] for x in range(0, len(posts_df_60), 100)]

increments_60 = get_lists(chunks)

get_posts(increments_60)

scores_df_60 = pd.DataFrame([ids, scores, comments]).T

scores_df_60.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_60.to_csv('./Data/5_to_6k_scores.csv', index=False)

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_70 = posts_df[6000:7000]

chunks = [posts_df_70['id'][x:x+100] for x in range(0, len(posts_df_70), 100)]

increments_70 = get_lists(chunks)

get_posts(increments_70)

scores_df_70 = pd.DataFrame([ids, scores, comments]).T

scores_df_70.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_70.to_csv('./Data/7_to_7k_scores.csv', index=False)

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_80 = posts_df[7000:8000]

chunks = [posts_df_80['id'][x:x+100] for x in range(0, len(posts_df_80), 100)]

increments_80 = get_lists(chunks)

get_posts(increments_80)

scores_df_80 = pd.DataFrame([ids, scores, comments]).T

scores_df_80.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_80.to_csv('./Data/7_to_8k_scores.csv', index=False)

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_90 = posts_df[8000:9000]

chunks = [posts_df_90['id'][x:x+100] for x in range(0, len(posts_df_90), 100)]

increments_90 = get_lists(chunks)

get_posts(increments_90)

scores_df_90 = pd.DataFrame([ids, scores, comments]).T

scores_df_90.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_90.to_csv('./Data/8_to_9k_scores.csv', index=False)

ids = []
scores = []
comments = []
def get_posts(items):
    for i in range(0, len(items)):
        posts = items[i]
        for submission in range(0,len(posts)):
            ids.append(posts[submission])
            scores.append(reddit.submission(posts[submission]).score)
            comments.append(reddit.submission(posts[submission]).num_comments)
    return ids, scores, comments

posts_df_100 = posts_df[9000:10001]

chunks = [posts_df_100['id'][x:x+100] for x in range(0, len(posts_df_100), 100)]

increments_100 = get_lists(chunks)

get_posts(increments_100)

scores_df_100 = pd.DataFrame([ids, scores, comments]).T

scores_df_100.rename(columns={0:'id', 1:'score', 2:'comments'}, inplace=True)
scores_df_100.to_csv('./Data/9_to_10k_scores.csv', index=False)
