==> 03- Preprocessing.ipynb <==
#### In this notebook I will pre-process the data to get it ready for modeling. 

import pandas as pd
import datetime as dt
import numpy as np
import pickle
import pytz 

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.compose import make_column_transformer, ColumnTransformer, make_column_selector
from sklearn.neighbors import KNeighborsRegressor

posts_df = pd.read_csv('../food_trends/Data/reddit_6_months.csv')

posts_df.drop(columns=['subreddit','selftext'], inplace=True)
posts_df.head(5)

#### first, I will check for any null values

posts_df.isnull().sum()

#### convert epoch time to datetime and convert UTC to EST 

resource used - https://www.adamsmith.haus/python/answers/how-to-convert-a-utc-datetime-to-a-local-datetime-in-python

date = dt.datetime.fromtimestamp(1653077037, pytz.utc)
date

local_datetime = pytz.timezone('US/Eastern')

date = date.astimezone(pytz.timezone('US/Eastern')).strftime("%Y-%m-%d, %H:%M")

date

posts_df['post_time_est']= [(dt.datetime.fromtimestamp(x, pytz.utc)).astimezone(pytz.timezone('US/Eastern')).strftime("%Y-%m-%d, %H:%M")
                            for x in posts_df['created_utc']]

posts_df['post_time_est'] = pd.to_datetime(posts_df['post_time_est'])

#### combine scores data with post data

posts_df.rename(columns={'score':'score_old', 'num_comments':'comments_old'}, inplace=True)

read in scores and comments from api

df_10 = pd.read_csv('./Data/0_to_10k_scores.csv')

df_20 = pd.read_csv('./Data/10_to_20k_scores.csv')

df_30 = pd.read_csv('./Data/2_to_3k_scores.csv')

df_40 = pd.read_csv('./Data/3_to_4k_scores.csv')

df_50 = pd.read_csv('./Data/4_to_5k_scores.csv')

df_60 = pd.read_csv('./Data/5_to_6k_scores.csv')

df_70 = pd.read_csv('./Data/7_to_7k_scores.csv')

df_80 = pd.read_csv('./Data/7_to_8k_scores.csv')

df_90 = pd.read_csv('./Data/8_to_9k_scores.csv')

df_100 = pd.read_csv('./Data/9_to_10k_scores.csv')

concatenate into 1 dataframe

scores_df = pd.concat([df_10, df_20, df_30, df_40, df_50, df_60, df_70, df_80, df_90, df_100])

scores_df.shape, posts_df.shape

scores_df.dtypes, posts_df.dtypes

scores_df.isnull().sum()

scores_df.dropna(inplace=True)

scores_df.drop_duplicates(inplace=True)

scores_df.shape

join to the posts dataframe

posts_scores_df = posts_df.merge(right=scores_df, how='inner')
posts_scores_df.head(5)

posts_scores_df.dtypes

posts_scores_df.isnull().sum()

posts_scores_df.shape

#### Engineer a field to connote 'viral'. Viral is being defined as a post with a score in the top 2% of all posts

posts_scores_df['percentile']= posts_scores_df.score.rank(pct=True)
#https://www.geeksforgeeks.org/percentile-rank-of-a-column-in-a-pandas-dataframe/

posts_scores_df['viral'] = np.where(posts_scores_df['percentile'] > .98, 1, 0 )

posts_scores_df.head()

posts_scores_df.to_csv('./Data/posts_scores_dates.csv', index=False)
