==> 05-01- Modeling (RNN).ipynb <==
This notebook is exploring applying Recurrent Neural Nets for the title data.

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split

from keras.preprocessing.text import Tokenizer
from keras.models import Sequential
from keras.layers import SimpleRNN, Dense, Embedding, Flatten, LSTM
from keras.preprocessing.sequence import pad_sequences

posts_df = pd.read_csv('./posts_scores_dates.csv')
posts_df.head()

posts_df.shape

posts_df.dtypes

X = posts_df['title']
y = posts_df['viral']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state = 20, stratify = y)

tokenizer = Tokenizer(num_words=500)

tokenizer.fit_on_texts(X_train)

X_train.head()

X_train = tokenizer.texts_to_sequences(X_train)
X_test = tokenizer.texts_to_sequences(X_test)

X_train = pad_sequences(X_train, maxlen=100)
X_test = pad_sequences(X_test, maxlen=100)

Basic RNN model

model = Sequential()
model.add(Embedding(input_dim=tokenizer.num_words, output_dim=64))
model.add(SimpleRNN(64))
model.add(Dense(64, activation='relu'))
model.add(Dense(1, activation='sigmoid'))

model.compile(loss='binary_crossentropy', metrics=['acc', 'bce'])

model.fit(X_train, y_train, validation_data=(X_test, y_test), epochs=10)

Basic LSTM model

model = Sequential()
model.add(Embedding(input_dim=tokenizer.num_words, output_dim=64))
model.add(LSTM(64))
model.add(Dense(64, activation='relu'))
model.add(Dense(1, activation='sigmoid'))

model.compile(loss='bce', metrics=['acc'])
model.fit(X_train, y_train, validation_data=(X_test, y_test), epochs=10)
