==> 5-02- Modeling (spacy_tokenizer).ipynb <==
%pip install -U spacy

import pandas as pd
import spacy
from sklearn.model_selection  import train_test_split
from spacy.tokens import DocBin

import spacy.cli

spacy.cli.download("en_core_web_md") 

posts_df = pd.read_csv('./posts_scores_dates.csv')
posts_df.head()

nlp = spacy.load('en_core_web_md')
nlp.pipeline

Try using SPACY as a tokenizer in sklearn countvectorizer

import numpy as np
from sklearn.neighbors import KNeighborsRegressor, KNeighborsClassifier
from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestRegressor, ExtraTreesRegressor, RandomForestClassifier
from sklearn.metrics import balanced_accuracy_score, SCORERS, mean_squared_error, accuracy_score, f1_score
from imblearn.over_sampling import RandomOverSampler
from imblearn.pipeline import make_pipeline, Pipeline

X = posts_df['title']
y = posts_df['viral']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=20)

def get_metrics(function):
    print(f'The training balanced accuracy score is {balanced_accuracy_score(y_train, function.predict(X_train))}')
    print(f'The testing balanced accuracy score is {balanced_accuracy_score(y_test, function.predict(X_test))}')
    print(f'The training f1 score is {f1_score(y_train, function.predict(X_train))}')
    print(f'The testing f1 score is {f1_score(y_test, function.predict(X_test))}')
    print(f'The training accuracy score is {accuracy_score(y_train, function.predict(X_train))}')
    print(f'The testing accuracy score is {accuracy_score(y_test, function.predict(X_test))}')

pipe_lr = Pipeline([
    ('cv', CountVectorizer()),
    ('lr', LogisticRegression(max_iter=1000, random_state=20, class_weight='balanced'))
])

pipe_lr.fit(X_train, y_train)

get_metrics(pipe_lr)

Create a tokenizer to use spacy tokenization

def spacy_tokenizer(doc):
  return [x.orth_ for x in nlp(doc)]

Test out the spacy tokenizer to make sure it works

cvs = CountVectorizer(tokenizer=spacy_tokenizer)

X_train_cvs = cvs.fit_transform(X_train)

pd.DataFrame(X_train_cvs.A, columns = cvs.get_feature_names_out()).head()

Now that the tokenizer is working, let's apply it to the best performing model from the other notebook

pipe_ros = Pipeline([
    ('cv', CountVectorizer()),
  ('ros', RandomOverSampler()),
    ('rfc', RandomForestClassifier())
])

params_ros  = {
    'cv__ngram_range': [(1,2)],
    'cv__min_df':[1, 2, 4],
    'cv__lowercase':[True, False],
    'cv__tokenizer' :[None, spacy_tokenizer],
    'ros__sampling_strategy':[0.0, .25, .5, .75, 1],
    'rfc__n_jobs':[-1],
}

gs_ros = GridSearchCV(
    pipe_ros,
    param_grid= params_ros,
    n_jobs=-1,
    scoring='balanced_accuracy'
)

gs_ros.fit(X_train, y_train)

gs_ros.best_estimator_

get_metrics(gs_ros)

The Random Oversampling and spacy tokenizer model times out on Google Colab and therefore could not complete fitting.
